Moving utility to the separate file
