+++ web/client/utils/EpicsUtils.js
@@ -6,14 +6,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {setControlProperty} from "../actions/controls";
-import Rx from "rxjs";
-import {OPEN_FEATURE_GRID} from "../actions/featuregrid";
-import {createControlEnabledSelector} from "../selectors/controls";
-import {REGISTER_EVENT_LISTENER} from "../actions/map";
-import {CHANGE_DRAWING_STATUS} from "../actions/draw";
-import {START_DRAWING} from "../actions/annotations";
-
 /**
  * Default wrapper for the epics. This avoids to close all epics system for unhandled exceptions.
  * It allows also to identify the error showing in console the name of the epic that triggered the exception and the error.
@@ -44,41 +36,3 @@ const defaultEpicWrapper = (epic, k = "--unknown--") => (...args) =>
 export const wrapEpics = (epics, wrapper = defaultEpicWrapper) =>
     Object.keys(epics).map(k => wrapper(epics[k], k));
 
-
-/**
- * Common part of the workflow that toggles one plugin off when another plugin intend to perform drawing action
- * @param action$
- * @param store
- * @param {string} toolName - tool name(s) that should be toggled off
- * @param {function} apply - optional function to override action triggered by default
- * @param isActiveCallback - optional function to override callback to check tool activeness
- * @returns {Observable<unknown>}
- */
-export const shutdownToolOnAnotherToolDrawing = (action$, store, toolName,
-    apply = (state, tool) => {
-        let actions = [
-            setControlProperty(tool, "enabled", null)
-        ];
-        return Rx.Observable.from(actions);
-    },
-    isActiveCallback = (state, name) => createControlEnabledSelector(name)(state)
-) =>
-    action$.ofType(START_DRAWING, CHANGE_DRAWING_STATUS, REGISTER_EVENT_LISTENER, OPEN_FEATURE_GRID)
-        .filter(({type, status, owner, eventName, toolName: name}) => {
-            const isActive = isActiveCallback(store.getState(), toolName);
-            switch (type) {
-            case OPEN_FEATURE_GRID:
-                return toolName !== 'featureGrid';
-            case REGISTER_EVENT_LISTENER:
-                return isActive && eventName === 'click' && name !== toolName;
-            case CHANGE_DRAWING_STATUS:
-                return isActive &&
-                    (status === 'drawOrEdit' || status === 'start') && owner !== toolName;
-            case START_DRAWING:
-            default:
-                return isActive && toolName !== 'annotations';
-            }
-        })
-        .switchMap((action) => {
-            return isActiveCallback(store.getState(), toolName) ? apply(store.getState(), toolName, action) : Rx.Observable.empty();
-        });
